refactor: 바텀시트 관련 패키지 정리
